% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing_risks_methods.R
\name{cumhaz2ts}
\alias{cumhaz2ts}
\title{Cumulated hazard over two time scales}
\usage{
cumhaz2ts(fitted_model, plot_grid = NULL, cause = NULL)
}
\arguments{
\item{fitted_model}{The output of the function \code{fit2ts}. This is an object of
class \code{'haz2ts'}.}

\item{plot_grid}{(optional) A list containing the parameters to build a new
finer grid of intervals over \code{u} and \code{s} for plotting. This must be of the
form:
\itemize{
\item \code{plot_grid = list(c(umin, umax, du), c(smin, smax, ds))}
where \code{umin}, \code{umax} and \code{smin}, \code{smax} are the minimum and maximum values
desired for the intervals over \code{u} and \code{s} respectively, and \code{du}, \code{ds} are
distances between intervals over \code{u} and \code{s} respectively. Specifying a new
denser grid is used to evaluate the B-spline bases used for estimation on
such grid and plot the estimated surfaces with a greater level of details.
If not specified, the plotting is done using the same B-splines bases as
for the estimation. The function will check if the parameters for the grid
provided by the user are compatible with those originally used to construct
the B-splines for estimating the model. If not, the grid will be adjusted
accordingly and a warning will be returned.
}}

\item{cause}{a character string with a short name for the cause (optional).}
}
\value{
A list with the following elements:
* \code{Haz} the 2d hazard surface (obtained from the function \code{get_hazard_2d});
* \code{CumHaz} the cumulated hazard estimate over \code{u} and \code{s};
* \code{cause} (if provided) the short name for the cause
}
\description{
Computes the cumulated hazard surface over two time scales
(currently only implemented for objects of class \code{haz2ts}, not \code{haz2tsLMM}).
}
