% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposures_events_Lexis.R
\name{exposures_events_Lexis}
\alias{exposures_events_Lexis}
\title{Bin data on the Lexis diagram}
\usage{
exposures_events_Lexis(t_in = NULL, t_out, s_in = NULL, s_out, ev, bins_list)
}
\arguments{
\item{t_in}{(optional) A vector of entry times on the time scale \code{t}.}

\item{t_out}{(optional) A vector of exit times on the time scale \code{t}.}

\item{s_in}{(optional) A vector of entry times on the time scale \code{s}.}

\item{s_out}{A vector of exit times on the time scale \code{s}.}

\item{ev}{A vector of event's indicators (possible values 0/1).}

\item{bins_list}{A list with the following (necessary) elements:
\itemize{
\item \code{bins_t} a vector of extreme values for the bins over the \code{t} axis.
\item \code{nt} the number of bins over \code{t}.
\item \code{bins_s} a vector of extreme values for the bins over the \code{t} axis.
\item \code{ns} the number of bins over \code{s}.
}}
}
\value{
A list with the following elements:
\itemize{
\item \code{R} an array of exposure times of dimension nt by ns
\item \code{Y} an array of event counts of dimension nt by ns
}
}
\description{
\code{exposures_events_Lexis()} computes aggregated matrices of exposure times and event counts over two time
scales, on the Lexis diagram.

The time scales are \code{t} and \code{s}. This function uses several functions of
the package \code{popEpi} and of the package Epi. See also \code{\link[=prepare_data]{prepare_data()}} to conveniently prepare individual data
for the analysis with one, or two time scales.
}
