% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2ts.R
\name{fit2ts}
\alias{fit2ts}
\title{Fit a smooth hazard model with two time scales}
\usage{
fit2ts(
  data2ts = NULL,
  Y = NULL,
  R = NULL,
  Z = NULL,
  bins = NULL,
  Bbases_spec = list(),
  pord = 2,
  optim_method = c("ucminf", "grid_search"),
  optim_criterion = c("aic", "bic"),
  lrho = c(0, 0),
  Wprior = NULL,
  ridge = 0,
  control_algorithm = list(),
  par_gridsearch = list()
)
}
\arguments{
\item{data2ts}{(optional) an object of class created by the function
\code{prepare_data()}. Proving this input is the easiest way to use the function
\code{fit2ts}. However, the user can also provide the input data together with
a list of bins, as explained by the following parameters' descriptions.}

\item{Y}{A matrix (or 3d-array) of event counts of dimension nu by ns
(or nu by ns by n).}

\item{R}{A matrix (or 3d-array) of exposure times of dimension nu by ns
(or nu by ns by n).}

\item{Z}{(optional) A regression matrix of covariates values of dimensions
n by p.}

\item{bins}{a list with the specification for the bins. This is created by
the function \code{prepare_data}. If a list prepared externally from such function
if provided, it should contain the following elements:
* \code{bins_u} A vector of bins extremes for the time scale \code{u}.
* \code{midu} A vector with the midpoints of the bins over \code{u}.
* \code{nu} The number of bins over \code{u}.
* \code{bins_s} A vector of bins extremes for the time scale \code{s}.
* \code{mids} A vector with the midpoints of the bins over \code{s}.
* \code{ns} The number of bins over \code{s}.}

\item{Bbases_spec}{A list with the specification for the B-splines basis
with the following elements:
\itemize{
\item \code{bdeg} The degree of the B-splines basis. Default is 3 (for cubic B-splines).
\item \code{nseg_u} The number of segments for the B-splines over \code{u}. Default is 10.
\item \code{min_u} (optional) The lower limit of the domain of \code{Bu}.
Default is \code{min(bins_u)}.
\item \code{max_u} (optional) The upper limit of the domain of \code{Bu}.
Default is \code{max(bins_u)}.
\item \code{nseg_s} The number of segments for the B-splines over \code{s}. Default is 10.
\item \code{min_s} (optional) The lower limit of the domain of \code{Bs}.
Default is \code{min(bins_s)}.
\item \code{max_s} (optional) The upper limit of the domain of \code{Bs}.
Default is \code{max(bins_s)}.
}}

\item{pord}{The order of the penalty. Default is 2.}

\item{optim_method}{The method to be used for optimization:
\code{"ucminf"} (default) for the numerical optimization of the AIC (or BIC)
and \code{"grid_search"} for a grid search of the minimum AIC (or BIC)
over a grid of \code{log_10(rho_u)} and \code{log_10(rho_s)} values.}

\item{optim_criterion}{The criterion to be used for optimization:
\code{"aic"} (default) or \code{"bic"}.}

\item{lrho}{A vector of two elements if \code{optim_method == "ucminf"}.
Default is \code{c(0,0)}. A list of two vectors of values for \code{log_10(rho_u)}
and \code{log_10(rho_s)} if \code{optim_method == "grid_search"}. In the latter case,
if a list with two vectors is not provided, a default sequence of
values is used for both \code{log_10(rho_u)} and \code{log_10(rho_s)}.}

\item{Wprior}{An optional matrix of a-priori weights.}

\item{ridge}{A ridge penalty parameter: default is 0.}

\item{control_algorithm}{A list with optional values for the parameters of
the iterative processes:
\itemize{
\item \code{maxiter} The maximum number of iteration for the IWSL algorithm.
Default is 20.
\item \code{conv_crit} The convergence criteria, expressed as difference between
estimates at iteration i and i+1. Default is \code{1e-5}.
\item \code{verbose} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the iteration
process.
\item \code{monitor_ev} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the
evaluation of the model over the \code{log_10(rho_s)} values.
}}

\item{par_gridsearch}{A list of parameters for the grid_search:
\itemize{
\item \code{plot_aic} A Boolean. Default is \code{FALSE}. If \code{TRUE}, plot the AIC values
over the grid of \code{log_10(rho_u)} and \code{log_10(rho_s)} values.
\item \code{plot_bic} A Boolean. Default is \code{FALSE}. If \code{TRUE}, plot the BIC values
over the grid of \code{log_10(rho_u)} and \code{log_10(rho_s)} values.
\item \code{return_aic} A Boolean. Default is \code{TRUE}. Return the AIC values.
\item \code{return_bic} A Boolean. Default is \code{TRUE}. Return the BIC values.
\item \code{col} The color palette to be used for the AIC/BIC plot. Default is
\code{grDevices::gray.colors(n=10)}.
\item \code{plot_contour} A Boolean. Default is \code{TRUE}. Adds white contour lines to
the AIC/BIC plot.
\item \code{mark_optimal} A Boolean. Default is \code{TRUE}. If the plot of the AIC or BIC
values is returned, marks the optimal combination of \code{log_10(rho_u)} and
\code{log_10(rho_s)} in the plot.
\item \code{main_aic} The title of the AIC plot. Default is \code{"AIC grid"}.
\item \code{main_bic} The title of the BIC plot. Default is \code{"BIC grid"}.
}}
}
\value{
An object of class \code{h2tsfit} with the following elements:
\itemize{
\item \code{optimal_model} A list with :
\itemize{
\item \code{Alpha} The matrix of estimated P-splines coefficients of dimension
cu by cs.
\item \code{Cov_alpha} The variance-covariance matrix of the \code{Alpha} coefficients,
of dimension cucs by cucs.
\item \code{beta} The vector of length p of estimated covariates coefficients
(if model with covariates).
\item \code{Cov_beta} The variance-covariance matrix of the \code{beta} coefficients,
of dimension p by p (if model with covariates).
\item \code{SE_beta} The vector of length p of estimated Standard Errors for the \code{beta}
coefficients (if model with covariates)..
\item \code{Eta} or \code{Eta0} The matrix of values of the (baseline) linear predictor
(log-hazard) of dimension nu by ns.
\item \code{H} The hat-matrix.
\item \code{deviance} The deviance.
\item \code{ed} The effective dimension of the model.
\item \code{aic} The value of the AIC.
\item \code{bic} The value of the BIC.
\item \code{Bbases} a list with the B-spline bases \code{Bu} and \code{Bs}
}
\item \code{optimal_logrho} A vector with the optimal values of \code{log10(rho_u)} and
\code{log10(rho_s)}.
\item \code{P_optimal} The optimal penalty matrix P.
\item \code{AIC} (if \code{par_gridsearch$return_aic == TRUE}) The matrix of AIC values.
\item \code{BIC} (if \code{par_gridsearch$return_bic == TRUE}) The matrix of BIC values.
}
}
\description{
\code{fit2ts()} fits a smooth hazard model with two time scales.

Two methods are implemented for the search of the optimal smoothing
parameters (and therefore optimal model): a numerical optimization of the
AIC or BIC of the model and a search for the minimum AIC or BIC of the
model over a grid of \code{log_10} values for the smoothing parameters.
Construction of the B-splines bases and of the penalty matrix is
incorporated within the function. If a matrix of covariates is provided,
the function will estimate a model with covariates.
}
