% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpgam.R
\name{fitpgam}
\alias{fitpgam}
\title{Fit a log-additive model over two time scales}
\usage{
fitpgam(
  data2ts = NULL,
  Y = NULL,
  R = NULL,
  Z = NULL,
  bins = NULL,
  Bbases_spec = list(),
  pord = 2,
  ridge = 1e-06,
  lrho = c(0, 0),
  optim_method = c("ucminf", "grid_search"),
  optim_criterion = c("aic", "bic"),
  control_algorithm = list(),
  par_gridsearch = list()
)
}
\arguments{
\item{data2ts}{(optional) an object of class \code{"data2ts"} created by the function
\code{prepare_data()}. Proving this input is the easiest way to use the function
\code{fitpgam()}. However, the user can also provide the input data together with
a list of bins, as explained by the following parameters' descriptions.}

\item{Y}{A matrix (or 3d-array) of event counts of dimension nu by ns
(or nu by ns by n).}

\item{R}{A matrix (or 3d-array) of exposure times of dimension nu by ns
(or nu by ns by n).}

\item{Z}{(optional) A regression matrix of covariates values of dimensions
n by p.}

\item{bins}{a list with the specification for the bins. This is created by
the function \code{prepare_data()}. If a list prepared externally from such function
if provided, it should contain the following elements:
* \code{bins_u} A vector of bins extremes for the time scale \code{u}.
* \code{midu} A vector with the midpoints of the bins over \code{u}.
* \code{nu} The number of bins over \code{u}.
* \code{bins_s} A vector of bins extremes for the time scale \code{s}.
* \code{mids} A vector with the midpoints of the bins over \code{s}.
* \code{ns} The number of bins over \code{s}.}

\item{Bbases_spec}{A list with the specification for the B-splines basis
with the following elements:
\itemize{
\item \code{bdeg} The degree of the B-splines basis. Default is 3 (for cubic B-splines).
\item \code{nseg_u} The number of segments for the B-splines over \code{u}. Default is 10.
\item \code{min_u} (optional) The lower limit of the domain of \code{Bu}.
Default is \code{min(bins_u)}.
\item \code{max_u} (optional) The upper limit of the domain of \code{Bu}.
Default is \code{max(bins_u)}.
\item \code{nseg_s} The number of segments for the B-splines over \code{s}. Default is 10.
\item \code{min_s} (optional) The lower limit of the domain of \code{Bs}.
Default is \code{min(bins_s)}.
\item \code{max_s} (optional) The upper limit of the domain of \code{Bs}.
Default is \code{max(bins_s)}.
}}

\item{pord}{The order of the penalty. Default is 2.}

\item{ridge}{A ridge penalty parameter: default is 0. This is useful when, in
some cases the algorithm shows convergence problems. In this case, set to a small
number, for example \code{1e-4}.}

\item{lrho}{A vector of two elements if \code{optim_method == "ucminf"}.
Default is \code{c(0,0)}. A list of two vectors of values for \eqn{\log_{10}(\rho_u)}
and \eqn{\log_{10}(\rho_s)} if \code{optim_method == "grid_search"}. In the latter case,
if a list with two vectors is not provided, a default sequence of
values is used for both \eqn{\log_{10}(\rho_u)} and \eqn{\log_{10}(\rho_s)}.}

\item{optim_method}{The method to be used for optimization:
\code{"ucminf"} (default) for the numerical optimization of the AIC (or BIC),
\code{"grid_search"} for a grid search of the minimum AIC (or BIC)
over a grid of \eqn{\log_{10}(\rho_u)} and \eqn{\log_{10}(\rho_s)} values.}

\item{optim_criterion}{The criterion to be used for optimization:
\code{"aic"} (default) or \code{"bic"}. BIC penalizes model complexity more strongly
than AIC, so that its usage is recommended when a smoother fit is preferable
(see also Camarda, 2012).}

\item{control_algorithm}{A list with optional values for the parameters of
the iterative processes:
\itemize{
\item \code{maxiter} The maximum number of iteration for the IWSL algorithm.
Default is 20.
\item \code{conv_crit} The convergence criteria, expressed as difference between
estimates at iteration i and i+1. Default is \code{1e-5}.
\item \code{verbose} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the iteration
process.
\item \code{monitor_ev} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the
evaluation of the model over the \code{log_10(rho_s)} values.
}}

\item{par_gridsearch}{A list of parameters for the grid_search:
\itemize{
\item \code{plot_aic} A Boolean. Default is \code{FALSE}. If \code{TRUE}, plot the AIC values
over the grid of \code{log_10(rho_u)} and \code{log_10(rho_s)} values.
\item \code{plot_bic} A Boolean. Default is \code{FALSE}. If \code{TRUE}, plot the BIC values
over the grid of \code{log_10(rho_u)} and \code{log_10(rho_s)} values.
\item \code{return_aic} A Boolean. Default is \code{TRUE}. Return the AIC values.
\item \code{return_bic} A Boolean. Default is \code{TRUE}. Return the BIC values.
\item \code{col} The color palette to be used for the AIC/BIC plot. Default is
\code{grDevices::gray.colors(n=10)}.
\item \code{plot_contour} A Boolean. Default is \code{TRUE}. Adds white contour lines to
the AIC/BIC plot.
\item \code{mark_optimal} A Boolean. Default is \code{TRUE}. If the plot of the AIC or BIC
values is returned, marks the optimal combination of \code{log_10(rho_u)} and
\code{log_10(rho_s)} in the plot.
\item \code{main_aic} The title of the AIC plot. Default is \code{"AIC grid"}.
\item \code{main_bic} The title of the BIC plot. Default is \code{"BIC grid"}.
}}
}
\value{
An object of class \code{haz2tsPGAM}, that is a list with the following elements:
\itemize{
\item \code{optimal_model} A list with :
\itemize{
\item \code{alpha} The vector of estimated P-splines coefficients of length
\eqn{c_u + c_s}.
\item \code{SE_alpha} The vector of estimated standard errors of the \code{alpha} coefficients,
of dimension \eqn{c_u + c_s}.
\item \code{beta} (if covariates) The vector of length \eqn{p} of estimated covariates coefficients.
\item \code{se_beta} The vector of length \eqn{p} of estimated Standard Errors for the \code{beta}
coefficients.
\item \code{eta0} (if covariates) The vector of values of the baseline linear predictor (log-hazard).
\item \code{eta} (if covariates) The vector of values of the baseline linear predictor
(log-hazard) of length \eqn{n_u * n_s}.
\item \code{H} The hat-matrix.
\item \code{deviance} The deviance.
\item \code{ed} The (total) effective dimension of the model.
\item \code{aic} The value of the AIC.
\item \code{bic} The value of the BIC.
\item \code{Bbases} a list with the B-spline bases \code{Bu} and \code{Bs}
}
\item \code{optimal_logrho} A vector with the optimal values of \eqn{\log_{10}(\rho_u)} and
\eqn{\log_{10}(\rho_s)}.
\item \code{P_optimal} The optimal penalty matrix P.
\item \code{EDu} The effective dimensions along the \code{u} axis.
\item \code{EDs} The effective dimensions along the \code{s} axis.
\item \code{AIC} (if \code{par_gridsearch$return_aic == TRUE}) The matrix of AIC values.
\item \code{BIC} (if \code{par_gridsearch$return_bic == TRUE}) The matrix of BIC values.
}
}
\description{
Fits a log-additive model over two time scales. If covariates are included in
the model, \code{fitpgam()} will fit a proportional hazards model, where the baseline
hazard is a surface over the two time scales, where the effect of the time scales
is additive on the log-scale. The model is described in Carollo et al. (2025).
}
\details{
The name P-GAM follows from the way this class of models is named in
Eilers and Marx (2021) Practical Smoothing. The Joys of P-splines.
In Carollo et al. (2025) this model is referred to as log-additive.

@references Carollo, A., Putter, H., Eilers, P. H. C., & Gampe, J. (2025).
Analysis of Time-to-Event Data With Two Time Scales.
An Application to Transitions out of Cohabitation.
Sociological Methods & Research, 0(0). \doi{10.1177/00491241251374193}
}
\examples{
# Create some fake data - the bare minimum
id <- 1:20
u <- c(
  5.43, 3.25, 8.15, 5.53, 7.28, 6.61, 5.91, 4.94, 4.25, 3.86, 4.05, 6.86,
  4.94, 4.46, 2.14, 7.56, 5.55, 7.60, 6.46, 4.96
)
s <- c(
  0.44, 4.89, 0.92, 1.81, 2.02, 1.55, 3.16, 6.36, 0.66, 2.02, 1.22, 3.96,
  7.07, 2.91, 3.38, 2.36, 1.74, 0.06, 5.76, 3.00
)
ev <- c(1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1) #'

fakedata <- as.data.frame(cbind(id, u, s, ev))
fakedata2ts <- prepare_data(
  data = fakedata,
  u = "u",
  s_out = "s",
  ev = "ev",
  ds = .5
)
# Fit a fake model - not optimal smoothing
fitpgam(fakedata2ts,
  optim_method = "grid_search",
  lrho = list(seq(1, 1.5, .5), seq(1, 1.5, .5))
)
}
