% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgam_functions.R
\name{fitpgammodel_ucminf}
\alias{fitpgammodel_ucminf}
\title{Numerical optimization of the P-GAM model with two time scales}
\usage{
fitpgammodel_ucminf(
  y,
  r,
  Z = NULL,
  optim_criterion = c("aic", "bic"),
  lrho = c(0, 0),
  B,
  nb,
  nbu,
  nbs,
  DutDu,
  DstDs,
  ridge = 1e-06,
  control_algorithm = list()
)
}
\arguments{
\item{Z}{(optional) A regression matrix of covariates values of dimensions
n by p.}

\item{optim_criterion}{The criterion to be used for optimization:
\code{"aic"} (default) or \code{"bic"}.}

\item{lrho}{A vector of two elements, the initial values for \eqn{\log_{10}(\varrho_u)}
and \eqn{\log_{10}(\varrho_s)}.}

\item{ridge}{A ridge penalty parameter: default is 0. This is useful when, in
some cases the algorithm shows convergence problems. In this case, set to a small
number, for example \code{1e-4}.}

\item{control_algorithm}{A list with optional values for the parameters of
the iterative processes:
\itemize{
\item \code{maxiter} The maximum number of iteration for the IWSL algorithm.
Default is 20.
\item \code{conv_crit} The convergence criteria, expressed as difference between
estimates at iteration i and i+1. Default is \code{1e-5}.
\item \code{verbose} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the iteration
process.
\item \code{monitor_ev} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the
evaluation of the model over the \code{log_10(rho_s)} values.
}}
}
\value{
An object of class \code{haz2tsPGAM} with the following elements:
\itemize{
\item \code{optimal_model} A list containing the results of the optimal model.
\item \code{optimal_logrho} A vector with the optimal values of \code{log10(rho_u)} and
\code{log10(rho_s)}.
\item \code{P_optimal} The optimal penalty matrix P.
}
}
\description{
\code{fitpgammodel_ucminf()} performs a numerical optimization of the
AIC or BIC of log-additive model with two time scales.

It finds the optimal values of \code{log_10(rho_u)} and \code{log_10(rho_s)}
and returns the estimated optimal model.
See also \code{\link[ucminf:ucminf]{ucminf::ucminf()}}.
}
\references{
Nielsen H, Mortensen S (2024).
\emph{ucminf: General-Purpose Unconstrained Non-Linear Optimization}.
R package version 1.2.2, \url{https://CRAN.R-project.org/package=ucminf}
}
\keyword{internal}
