% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitvcm.R
\name{fitvcm}
\alias{fitvcm}
\title{Fit a Varying Coefficient Model (VCM) over two time scales}
\usage{
fitvcm(
  data2ts = NULL,
  Y = NULL,
  R = NULL,
  Z = NULL,
  bins = NULL,
  Bbases_spec = list(),
  pord = 2,
  kappa = 1e-10,
  lsmpar = c(0, 0),
  control_algorithm = list()
)
}
\arguments{
\item{data2ts}{(optional) an object of class \code{"data2ts"} created by the function
\code{prepare_data()}. Proving this input is the easiest way to use the function
\code{fitvcm()}. However, the user can also provide the input data together with
a list of bins, as explained by the following parameters' descriptions.}

\item{Y}{A matrix (or 3d-array) of event counts of dimension nu by ns
(or nu by ns by n).}

\item{R}{A matrix (or 3d-array) of exposure times of dimension nu by ns
(or nu by ns by n).}

\item{Z}{(optional) A regression matrix of covariates values of dimensions
n by p.}

\item{bins}{a list with the specification for the bins. This is created by
the function \code{prepare_data()}. If a list prepared externally from such function
if provided, it should contain the following elements:
* \code{bins_u} A vector of bins extremes for the time scale \code{u}.
* \code{midu} A vector with the midpoints of the bins over \code{u}.
* \code{nu} The number of bins over \code{u}.
* \code{bins_s} A vector of bins extremes for the time scale \code{s}.
* \code{mids} A vector with the midpoints of the bins over \code{s}.
* \code{ns} The number of bins over \code{s}.}

\item{Bbases_spec}{A list with the specification for the B-splines basis
with the following elements:
\itemize{
\item \code{bdeg} The degree of the B-splines basis. Default is 3 (for cubic B-splines).
\item \code{nseg_s} The number of segments for the B-splines over \code{s}. Default is 10.
\item \code{min_s} (optional) The lower limit of the domain of \code{Bs}.
Default is \code{min(bins_s)}.
\item \code{max_s} (optional) The upper limit of the domain of \code{Bs}.
Default is \code{max(bins_s)}.
}}

\item{pord}{The order of the penalty. Default is 2.}

\item{kappa}{A ridge penalty.}

\item{lsmpar}{The starting values for the two smoothing parameters.
A vector of two elements, default is \code{c(0,0)}.}

\item{control_algorithm}{A list with optional values for the parameters of
the iterative processes:
\itemize{
\item \code{maxiter} The maximum number of iteration for the IWSL algorithm.
Default is 20.
\item \code{conv_crit} The convergence criteria, expressed as difference between
estimates at iteration i and i+1. Default is \code{1e-5}.
\item \code{verbose} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the iteration
process.
\item \code{monitor_ev} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the
evaluation of the model over the \code{log_10(rho_s)} values.
}}
}
\value{
An object of class \code{haz2tsVCM}, that is a list with the following elements:
\itemize{
\item \code{optimal_model} A list with :
\itemize{
\item \code{Alpha} The matrix of estimated P-splines coefficients of dimension
\eqn{c_s} by 2.
\item \code{Cov_alpha} The variance-covariance matrix of the \code{Alpha} coefficients,
of dimension \eqn{c_uc_s} by \eqn{c_uc_s}.
\item \code{Eta} The matrix of values of the baseline linear predictor
(log-hazard) of dimension \eqn{n_u} by \eqn{n_s}.
\item \code{H} The hat-matrix.
\item \code{deviance} The deviance.
\item \code{ed} The effective dimension of the model.
\item \code{aic} The value of the AIC.
\item \code{bic} The value of the BIC.
\item \code{Bbases} a list with the B-spline bases \code{Bu} and \code{Bs}
}
\item \code{optimal_logsmpar} A vector with the optimal values of \eqn{\log_{10}(\theta)} and \eqn{\log_{10}(\phi)}.
\item \code{P_optimal} The optimal penalty matrix P.
\item \code{AIC} (if \code{par_gridsearch$return_aic == TRUE}) The matrix of AIC values.
\item \code{BIC} (if \code{par_gridsearch$return_bic == TRUE}) The matrix of BIC values.
}

@references Carollo, A., Putter, H., Eilers, P. H. C., & Gampe, J. (2025).
Analysis of Time-to-Event Data With Two Time Scales.
An Application to Transitions out of Cohabitation.
Sociological Methods & Research, 0(0). \doi{10.1177/00491241251374193}
}
\description{
\code{fitvcm()} fits a varying coefficient model over two time scales with P-splines.
The model is described in Carollo et al. (2025).
}
\examples{
# Create some fake data - the bare minimum
id <- 1:20
u <- c(
  5.43, 3.25, 8.15, 5.53, 7.28, 6.61, 5.91, 4.94, 4.25, 3.86, 4.05, 6.86,
  4.94, 4.46, 2.14, 7.56, 5.55, 7.60, 6.46, 4.96
)
s <- c(
  0.44, 4.89, 0.92, 1.81, 2.02, 1.55, 3.16, 6.36, 0.66, 2.02, 1.22, 3.96,
  7.07, 2.91, 3.38, 2.36, 1.74, 0.06, 5.76, 3.00
)
ev <- c(1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1) #'

fakedata <- as.data.frame(cbind(id, u, s, ev))
fakedata2ts <- prepare_data(
  data = fakedata,
  u = "u",
  s_out = "s",
  ev = "ev",
  ds = .5
)
# Fit a fake model - not optimal smoothing
fitvcm(fakedata2ts)

}
