% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgam_functions.R
\name{get_aic_fit_pgam}
\alias{get_aic_fit_pgam}
\title{Return the AIC of P-GAM model}
\usage{
get_aic_fit_pgam(
  lrho,
  r,
  y,
  Z = NULL,
  B,
  nb,
  nbu,
  nbs,
  DutDu,
  DstDs,
  ridge = 1e-06,
  control_algorithm = list(maxiter = 20, conv_crit = 1e-05, verbose = FALSE, monitor_ev =
    FALSE)
)
}
\arguments{
\item{lrho}{A vector of two elements, the initial values for \eqn{\log_{10}(\varrho_u)}
and \eqn{\log_{10}(\varrho_s)}.}

\item{r}{A vector of exposure times of length \code{nu} by \code{ns}.}

\item{y}{A vector of event counts of length \code{nu} by \code{ns}.}

\item{Z}{(optional) A regression matrix of covariates values of dimensions
n by p.}

\item{B}{A matrix of B-splines.}

\item{nb}{The number of B-splines.}

\item{nbu}{The number of B-splines for the \code{u} dimension.}

\item{nbs}{The number of B-splines for the \code{s} dimension.}

\item{DutDu}{\eqn{D_u^TD_u} to multiply \eqn{\varrho_u}.}

\item{DstDs}{\eqn{D_s^TD_s} to multiply \eqn{\varrho_s}.}

\item{ridge}{A ridge penalty parameter: default is 0. This is useful when, in
some cases the algorithm shows convergence problems. In this case, set to a small
number, for example \code{1e-6}.}

\item{control_algorithm}{A list with optional values for the parameters of
the iterative processes:
\itemize{
\item \code{maxiter} The maximum number of iteration for the IWSL algorithm.
Default is 20.
\item \code{conv_crit} The convergence criteria, expressed as difference between
estimates at iteration i and i+1. Default is \code{1e-5}.
\item \code{verbose} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the iteration
process.
\item \code{monitor_ev} A Boolean. Default is \code{FALSE}. If \code{TRUE} monitors the
evaluation of the model over the \code{log_10(rho_s)} values.
}}
}
\value{
The \code{aic} value of the fitted model.
}
\description{
\code{get_aic_fit_pgam()} fits the log-additive model with two time scales
with or without individual level covariates and it returns the AIC of the model.
See also \code{fitpgammodel_ucminf()} and \code{fitpgam()}.
}
\keyword{internal}
