% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAIC_BIC_LMM.R
\name{getAIC_BIC_LMM}
\alias{getAIC_BIC_LMM}
\title{Calculates AIC and BIC from object fitted via LMMsolver}
\usage{
getAIC_BIC_LMM(fit, offset)
}
\arguments{
\item{fit}{An object of class \code{"LMMsolve"}}

\item{offset}{The vector of exposure times from dataLMM}
}
\value{
A list with ED, ED of the baseline hazard, deviance, AIC, BIC and
number of estimated covariate's parameters (if model with covariates)
}
\description{
Calculates AIC and BIC from object fitted via LMMsolver
}
