% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make a grid of points to evaluate B-splines}
\usage{
make_grid(
  plot_grid,
  model_specifications,
  class_fitmodel,
  where_slices = NULL,
  direction = c(NULL, "u", "s"),
  tmax = NULL,
  midpoints = FALSE
)
}
\arguments{
\item{plot_grid}{(optional) A list containing the parameters to build a new
finer grid of intervals over u and s for plotting. This must be of the
form: \code{plot_grid = list(c(umin, umax, du), c(smin, smax, ds))}, where
\code{umin}, \code{umax} and \code{smin}, \code{smax} are the minimum and maximum values
desired for the intervals over \code{u} and \code{s} respectively, and \code{du}, \code{ds} are
distances between intervals over \code{u} and \code{s} respectively. Specifying a new
denser grid is used to evaluate the B-spline bases used for estimation on
such grid and plot the estimated surfaces with a greater level of details.
If not specified, the plotting is done using the same B-splines bases as
for the estimation. The function will check if the parameters for the grid
provided by the user are compatible with those originally used to construct
the B-splines for estimating the model. If not, the grid will be adjusted
accordingly and a warning will be returned.}

\item{model_specifications}{A list with \code{'umin'}, \code{'umax'}, \code{'smin'}, and \code{'smax'} used for fitting.}

\item{class_fitmodel}{The class of the fitted model. Can be \code{'haz2ts'}, \code{'haz2tsPGAM'}, or \code{'haz2tsVCM'}.}

\item{where_slices}{A vector of values for the cutting points of the desired
slices of the surface. If \code{which_plot == "slices"}, please provide this
argument. Please also provide this argument in case \verb{which_plot = "survival}
or \verb{which_plot = "cumhaz} and \code{surv_slices = TRUE} or \code{cumhaz_slices = TRUE},
respectively.}

\item{direction}{If \code{which_plot == "slices"}, indicates the direction for
cutting the surface. If \code{u}, then the surface will be cut at the selected
values of \code{u} (indicated by \code{where_slices}), hence obtaining one-dimensional
curves over \code{s}. If \code{s}, then the surface will be cut at the selected values
of \code{s} (indicated by \code{where_slices}), hence obtaining one-dimensional curves
over \code{u}.}

\item{tmax}{The maximum value of \code{t} that should be plotted.}

\item{midpoints}{A Boolean. Default is \code{FALSE}. If \code{TRUE}, the estimated
quantities are evaluated at the midpoints of the rectangles
(or parallelograms) of the grids, rather than at each grid-point.}
}
\value{
A list of specification for the grid
}
\description{
Make a grid of points to evaluate B-splines
}
\examples{
make_grid(plot_grid = list(
    c(umin = 3, umax = 8.5, du = .1),
    c(smin = 0, smax = 7.1, ds = .1)
  ),
  model_specification = list(umin = 2, umax = 8.5, smin = 0, smax = 7.5),
  class_fitmodel = "haz2ts")

}
