% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_haz2ts.R
\name{predict_haz2ts}
\alias{predict_haz2ts}
\title{Prediction method for objects of class \code{'haz2ts'}}
\usage{
predict_haz2ts(
  x,
  newdata = NULL,
  originaldata = NULL,
  u,
  s,
  z = NULL,
  id = NULL,
  ds = NULL
)
}
\arguments{
\item{x}{an object of class \code{'haz2ts'}, the output of the function \code{fit2ts()}.}

\item{newdata}{(optional) A dataframe with columns cointaing the values of the variable \code{u}
and the time scale \code{s} for which predictions are to be obtained.}

\item{originaldata}{(optional) The original dataset. Provide it to obtain individual predictions
for each observation in the data.}

\item{u}{The name of the variable in \code{newdata}, or in \code{originaldata} containing
values for the variable \code{u}.}

\item{s}{The name of the variable in \code{newdata}, or in \code{originaldata} containing
values for the variable \code{s}. Note that over the \code{s} axis predictions
are provided only within intervals of values, as it is necessary to
approximate cumulated quantities.}

\item{z}{Covariates value}

\item{id}{(optional) The name of the variable in \code{newdata}, or in \code{originaldata} containing
the identification of each observation. It is not required for predictions on a
new dataset.}

\item{ds}{(optional) The distance between two consecutive points on the \code{s} axis.
If not provided, an optimal minimum value will be chosen automatically and
a warning is returned.}
}
\value{
A dataframe. This can be the original dataframe (\code{originaldata}), where only the
variables \code{id}, \code{u} and \code{s} are selected, or the new data frame (\code{newdata}),
together with the predicted values for the hazard \code{hazard} and its
standard errors \code{se_hazard}, the cumulative hazard \code{cumhazard} and the
survival probability \code{survival}.
}
\description{
Prediction method for objects of class \code{'haz2ts'}
}
\details{
Predictions of cumulated quantities can be provided only within intervals of values on the \code{s} time scale.
}
\examples{
# Create the same fake data as in other examples
id <- 1:20
u <- c(
  5.43, 3.25, 8.15, 5.53, 7.28, 6.61, 5.91, 4.94, 4.25, 3.86, 4.05, 6.86,
  4.94, 4.46, 2.14, 7.56, 5.55, 7.60, 6.46, 4.96
)
s <- c(
  0.44, 4.89, 0.92, 1.81, 2.02, 1.55, 3.16, 6.36, 0.66, 2.02, 1.22, 3.96,
  7.07, 2.91, 3.38, 2.36, 1.74, 0.06, 5.76, 3.00
)
ev <- c(1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1)
z1 <- rbinom(n = 20, size = 20, prob = .5)
z2 <- rnorm(n = 20)

fakedata <- as.data.frame(cbind(id, u, s, ev, z1, z2))
fakedata2ts <- prepare_data(
  u = fakedata$u,
  s_out = fakedata$s,
  ev = fakedata$ev,
  ds = .5, individual = TRUE,
  covs = subset(fakedata, select = c("z1", "z2"))
)

# Fit a fake model - not optimal smoothing
fakemod <- fit2ts(fakedata2ts,
  optim_method = "grid_search",
  lrho = list(
    seq(1, 1.5, .5),
    seq(1, 1.5, .5)
  )
)
# Create a new dataset for prediction
newdata <- as.data.frame(cbind("u" = c(2.5, 3.4, 6),
                               "s" = c(.2, .5, 1.3)))

# First - predict on original data
predict(object = fakemod,
  originaldata = fakedata, u = "u", s = "s", id = "id"
)


# Now - predict on new dataset
predict(object = fakemod,
  newdata = newdata, u = "u", s = "s"
)

# Now - predict including covariates
predict(object = fakemod,
        originaldata = fakedata,
        u = "u", s = "s", id = "id",
        z = c("z1", "z2")
)
# If one wants to predict with only one of the covariates at a different
# value than the baseline, the other one(s) should be fixed at their
# baseline levels

newdata2 <- as.data.frame(cbind("u" = c(2.5, 3.4, 6),
                                "s" = c(.2, .5, 1.3),
                                "z1" = c(1, 2, 3),
                                "z2" = c(0, 0, 0)))
predict(object = fakemod,
        newdata = newdata2,
        u = "u", s = "s", id = "id",
        z = c("z1", "z2")
)

}
