% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.data2ts.R
\name{print.data2ts}
\alias{print.data2ts}
\title{Print method for a \code{data2ts} object}
\usage{
\method{print}{data2ts}(x, ...)
}
\arguments{
\item{x}{of class \code{data2ts}, as prepared by \code{\link{prepare_data}}}

\item{\dots}{Further arguments to print}
}
\value{
No return value
}
\description{
Print method for an object of class \code{data2ts}
}
\examples{
# Bin the colon cancer data over s (time since recurrence)

dt1ts <- prepare_data(data = reccolon2ts,
                      s_in = "entrys",
                      s_out = "timesr",
                      events = "status",
                      ds = 180)

print(dt1ts)

# Bin the colon cancer data over u (time at recurrence) and s (time since recurrence)
dt2ts <- prepare_data(data = reccolon2ts,
                      u = "timer",
                      s_in = "entrys",
                      s_out = "timesr",
                      events = "status",
                      ds = 180)
print(dt2ts)

}
\author{
Angela Carollo \email{carollo@demogr.mpg.de}
}
