% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing_risks_methods.R
\name{surv2ts}
\alias{surv2ts}
\title{Survival functions (cause-specific) and overall survival with two time scales}
\usage{
surv2ts(cumhaz = list(), cause = NULL)
}
\arguments{
\item{cumhaz}{a list with all the cause-specific cumulated hazard matrices
(minimum 1). If more than one cause-specific cumulated hazard is provided,
then they should all be matrices of the same dimension.}

\item{cause}{is an optional vector of short names for the causes. It should
be of the same length as the number of cause-specific cumulated hazards provided.}
}
\value{
a list with cause-specific survival probability matrices and
one overall survival probability matrix (if at least two cause-specific
cumulated hazards are provided)
}
\description{
Computes cause-specific survival functions over two time scales from a list of
cause-specific cumulated hazard matrices. Each matrix of cause-specific survival
is calculated as exp(-CumHaz). Additionally, if the user provides more than
one cause-specific cumulated hazard, the function computes the overall survival
matrix, that contains the probability of not experiencing an event of any cause
by time \code{s} and fixed time at entry \code{u}.
}
