% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposures_events_1d.R
\name{exposures_events_1d}
\alias{exposures_events_1d}
\title{Bin data on one time scale}
\usage{
exposures_events_1d(s_in = NULL, s_out, ev, bins)
}
\arguments{
\item{s_in}{A vector of (possibly left truncated) times at entry. If this is
not provided by the user, the function will consider a value of 0 for all
observations.}

\item{s_out}{A vector of times at event or censoring.}

\item{ev}{A vector of event's indicators (possible values 0/1).}

\item{bins}{A vector of interval breaks for discretization.}
}
\value{
A list with the following elements:
\itemize{
\item \code{R} A matrix of dimension n by ns containing the exposure times for each
individual separately.
\item \code{r} A vector of exposure times.
\item \code{Y} A matrix of dimension n by ns containing the event counts for each
individual separately
\item \code{y} A vector of event counts.
}

If the length of the input vectors do not match, an error message is
returned.
}
\description{
\code{exposure_events_1d()} computes aggregated measures of exposure
times and event counts starting from individual records of time at entry,
time at exit and event's indicator.
}
\details{
The time scale \code{s} is divided into bins of equal size, which are
provided as input in the function. Then, the time-at-risk for each
individual is split according to these bins, and an event indicator is
placed in the bin where the exit time is located. Finally, the individual
contributions are summed in each bin to provide a vector of total exposure
time and total event counts in each bin. See also \code{\link[=prepare_data]{prepare_data()}} to
conveniently prepare individual data for the analysis with one, or two time
scales.
}
\examples{
# ---- Bin colon cancer data by time since recurrence ----
# First create vector of bins
K <- ceiling((max(reccolon2ts$timesr) - min(reccolon2ts$timesr)) / 30)
bins_s <- seq(min(reccolon2ts$timesr), min(reccolon2ts$timesr) + K * 30, by = 30)
# Now bin data (note: the s_in argument is omitted because data are not left truncated)
bindata <- exposures_events_1d(s_out = reccolon2ts$timesr, ev = reccolon2ts$status, bins = bins_s)

}
