% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hazard_2d.R
\name{get_hazard_2d}
\alias{get_hazard_2d}
\title{Get estimated (log-)hazard surface with 2 time scales}
\usage{
get_hazard_2d(
  fitted_model,
  plot_grid = NULL,
  where_slices = NULL,
  direction = c("u", "s", NULL)
)
}
\arguments{
\item{fitted_model}{The output of the function \code{fit2ts}. This is an object of
class \code{"h2tsfit"}.}

\item{plot_grid}{(optional) A list containing the parameters to build a new
finer grid of intervals over \code{u} and \code{s} for plotting. This must be of the
form: \code{plot_grid = list(c(umin, umax, du), c(smin, smax, ds))}, where
\code{umin}, \code{umax} and \code{smin}, \code{smax} are the minimum and maximum values
desired for the intervals over \code{u} and \code{s} respectively, and \code{du}, \code{ds} are
distances between intervals over \code{u} and \code{s} respectively. Specifying a new
denser grid is used to evaluate the B-spline bases used for estimation on
such grid and plot the estimated surfaces with a greater level of details.
If not specified, the plotting is done using the same B-splines bases as
for the estimation. The function will check if the parameters for the grid
provided by the user are compatible with those originally used to construct
the B-splines for estimating the model. If not, the grid will be adjusted
accordingly and a warning will be returned.}

\item{where_slices}{A vector of values for the cutting points of the desired
slices of the surface. If \code{which_plot == "slices"}, please provide this
argument.}

\item{direction}{If \code{which_plot == "slices"}, indicates the direction for
cutting the surface. If \code{u}, then the surface will be cut at the selected
values of \code{u} (indicated by \code{where_slices}), hence obtaining one-dimensional
curves over \code{s}. If \code{s}, then the surface will be cut at the seleced values
of \code{s} (indicated by \code{where_slices}), hence obtaining one-dimensional curves
over \code{u}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{new_plot_grid} A list of new specifications of the grid for plotting.
\item \code{hazard} A matrix containing the estimated hazard.
\item \code{loghazard} A matrix containing the estimated log-hazard.
\item \code{SE_hazard} A matrix containing the estimated SEs for the hazard
\item \code{SE_loghazard} A matrix containing the estimated SEs for the log-hazard
}
}
\description{
\code{get_hazard_2d()} takes as input the results of a model
estimated by \code{fit2ts} and it returns the estimated smooth log-hazard
and the smooth hazard together with their standard errors.

If the model includes covariates, then only the baseline (log-)hazard is returned.
It is possible to provide values that define a new grid for plotting.
If not specified, the plotting is done using the same B-splines basis as for
the estimation. The function will check if the parameters for the grid
provided by the user are compatible with those originally used to construct
the B-splines for estimating the model. If not, the grid will be adjusted
accordingly and a warning will be returned.
}
