% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haz2tsLMM_summary.R
\name{haz2tsLMM_summary}
\alias{haz2tsLMM_summary}
\title{Summary function for object of class 'haz2tsLMM'}
\usage{
haz2tsLMM_summary(x, ...)
}
\arguments{
\item{x}{an object of class 'haz2tsLMM' returned by the function \code{\link[=fit2ts]{fit2ts()}}}

\item{...}{further arguments}
}
\value{
a printed summary of the fitted model, including optimal smoothing
paramters, the effective dimension ED and the AIC/BIC. For model
with covariates, a regression table is also returned.
}
\description{
Summary function for object of class 'haz2tsLMM'
}
