% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competing_risks_methods.R
\name{surv2ts}
\alias{surv2ts}
\title{Survival function with two time scales}
\usage{
surv2ts(cumhaz = list())
}
\arguments{
\item{cumhaz}{a list with all the cause-specific cumulated hazard matrices
(minimum one element needs to be supplied).
If more than one cause-specific cumulated hazard is provided,
then they should all be matrices of the same dimension.}
}
\value{
a matrix containing the values of the survival function over \code{s} and \code{u}.
}
\description{
Computes the survival matrix, that contains the probability of not
experiencing an event of any cause by time \code{s} and fixed time at entry \code{u},
from a list of cause-specific cumulated hazard matrices.
}
